Multi-page file import methods available as instance methods on GiniVision
Issue #173
