+++ ginivision/src/main/java/net/gini/android/vision/GiniVisionFileImport.java
@@ -1,29 +1,23 @@
 package net.gini.android.vision;
 
-import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.os.AsyncTask;
 import android.support.annotation.NonNull;
 
 import net.gini.android.vision.analysis.AnalysisActivity;
 import net.gini.android.vision.document.DocumentFactory;
 import net.gini.android.vision.document.ImageDocument;
 import net.gini.android.vision.document.ImageMultiPageDocument;
-import net.gini.android.vision.internal.camera.photo.Photo;
-import net.gini.android.vision.internal.camera.photo.PhotoFactory;
 import net.gini.android.vision.internal.util.ActivityHelper;
 import net.gini.android.vision.internal.util.DeviceHelper;
 import net.gini.android.vision.internal.util.FileImportValidator;
-import net.gini.android.vision.internal.util.MimeType;
 import net.gini.android.vision.review.ReviewActivity;
 import net.gini.android.vision.review.multipage.MultiPageReviewActivity;
 import net.gini.android.vision.util.CancellationToken;
 import net.gini.android.vision.util.IntentHelper;
 import net.gini.android.vision.util.UriHelper;
 
-import java.io.IOException;
 import java.util.List;
 
 /**
@@ -31,6 +25,9 @@ import java.util.List;
  * from another app.
  */
 public final class GiniVisionFileImport {
+
+    @NonNull
+    private final GiniVision mGiniVision;
 
     /**
      * <b>Screen API</b>
@@ -139,7 +136,7 @@ public final class GiniVisionFileImport {
      * @param context               Android context subclass
      * @throws ImportedFileValidationException if the file didn't pass validation
      */
-    public static CancellationToken createIntentForImportedFiles(@NonNull final Intent intent,
+    CancellationToken createIntentForImportedFiles(@NonNull final Intent intent,
             @NonNull final Context context,
             @NonNull final Callback<Intent> callback) {
         final CancellationToken cancellationToken =
@@ -170,7 +167,7 @@ public final class GiniVisionFileImport {
     }
 
     @NonNull
-    private static Intent createIntent(final @NonNull Document result,
+    private Intent createIntent(final @NonNull Document result,
             final @NonNull Context context) {
         final Intent giniVisionIntent;
         if (result.getType() == Document.Type.IMAGE_MULTI_PAGE) {
@@ -195,17 +192,8 @@ public final class GiniVisionFileImport {
         return giniVisionIntent;
     }
 
-    public static CancellationToken createDocumentForImportedFiles(@NonNull final Intent intent,
+    CancellationToken createDocumentForImportedFiles(@NonNull final Intent intent,
             @NonNull final Context context, @NonNull final Callback<Document> callback) {
-        if (!GiniVision.hasInstance()) {
-            callback.onFailed(new ImportedFileValidationException(
-                    "Cannot import files. GiniVision instance not available. Create it with GiniVision.newInstance()."));
-            return new CancellationToken() {
-                @Override
-                public void cancel() {
-                }
-            };
-        }
         final List<Uri> uris = IntentHelper.getUris(intent);
         if (uris == null) {
             callback.onFailed(
@@ -217,15 +205,10 @@ public final class GiniVisionFileImport {
             };
         }
         final ImportFilesAsyncTask asyncTask = new ImportFilesAsyncTask(context, intent,
-                new Callback<ImageMultiPageDocument>() {
+                mGiniVision, new Callback<ImageMultiPageDocument>() {
                     @Override
                     public void onDone(@NonNull final ImageMultiPageDocument result) {
-                        if (!GiniVision.hasInstance()) {
-                            callback.onFailed(new ImportedFileValidationException(
-                                    "Cannot import files. GiniVision instance not available. Create it with GiniVision.newInstance()."));
-                            return;
-                        }
-                        GiniVision.getInstance().internal().getImageMultiPageDocumentMemoryStore()
+                        mGiniVision.internal().getImageMultiPageDocumentMemoryStore()
                                 .setMultiPageDocument(result);
                         callback.onDone(result);
                     }
@@ -249,7 +232,8 @@ public final class GiniVisionFileImport {
         };
     }
 
-    private GiniVisionFileImport() {
+    GiniVisionFileImport(@NonNull final GiniVision giniVision) {
+        mGiniVision = giniVision;
     }
 
     public interface Callback<T> {
@@ -261,112 +245,4 @@ public final class GiniVisionFileImport {
         void onCancelled();
     }
 
-    private static class ImportFilesAsyncTask extends AsyncTask<Uri, Void, ImageMultiPageDocument> {
-
-        @SuppressLint("StaticFieldLeak")
-        private final Context mContext;
-        private final Intent mIntent;
-        private final Callback<ImageMultiPageDocument> mCallback;
-        private ImportedFileValidationException mException;
-
-        private ImportFilesAsyncTask(@NonNull final Context context,
-                @NonNull final Intent intent,
-                @NonNull final Callback<ImageMultiPageDocument> callback) {
-            mContext = context;
-            mIntent = intent;
-            mCallback = callback;
-        }
-
-        @Override
-        protected ImageMultiPageDocument doInBackground(final Uri... uris) {
-            final ImageMultiPageDocument multiPageDocument = new ImageMultiPageDocument(
-                    Document.Source.newExternalSource(), Document.ImportMethod.OPEN_WITH);
-            for (final Uri uri : uris) {
-                if (isCancelled()) {
-                    return null;
-                }
-                if (!UriHelper.isUriInputStreamAvailable(uri, mContext)) {
-                    mException = new ImportedFileValidationException(
-                            "InputStream not available for one of the Intent's data Uris");
-                    return null;
-                }
-                final FileImportValidator fileImportValidator = new FileImportValidator(mContext);
-                if (fileImportValidator.matchesCriteria(uri)) {
-                    if (isCancelled()) {
-                        return null;
-                    }
-                    if (IntentHelper.hasMimeTypeWithPrefix(uri, mContext,
-                            MimeType.IMAGE_PREFIX.asString())) {
-                        final String deviceOrientation = DeviceHelper.getDeviceOrientation(
-                                mContext);
-                        final String deviceType = DeviceHelper.getDeviceType(mContext);
-                        // Create Document
-                        final ImageDocument document = DocumentFactory.newImageDocumentFromUri(uri,
-                                mIntent, mContext, deviceOrientation,
-                                deviceType, Document.ImportMethod.OPEN_WITH);
-                        try {
-                            // Load uri into memory
-                            final byte[] bytesFromUri = UriHelper.getBytesFromUri(uri, mContext);
-                            document.setData(bytesFromUri);
-                        } catch (final IOException e) {
-                            mException = new ImportedFileValidationException(
-                                    "Failed to read file into memory");
-                            return null;
-                        }
-                        if (isCancelled()) {
-                            return null;
-                        }
-                        // Create Photo
-                        final Photo photo = PhotoFactory.newPhotoFromDocument(document);
-                        if (isCancelled()) {
-                            return null;
-                        }
-                        // Compress Photo
-                        photo.edit().compressByDefault().apply();
-                        if (isCancelled()) {
-                            return null;
-                        }
-                        // Save to local storage
-                        final Uri localUri = GiniVision.getInstance().internal().getImageDiskStore()
-                                .save(mContext, photo.getData());
-                        if (localUri == null) {
-                            mException = new ImportedFileValidationException(
-                                    "Failed to copy to app storage");
-                            return null;
-                        }
-                        if (isCancelled()) {
-                            return null;
-                        }
-                        // Create compressed Document
-                        final ImageDocument compressedDocument =
-                                DocumentFactory.newImageDocumentFromPhoto(photo, localUri);
-                        multiPageDocument.addDocument(compressedDocument);
-                    }
-                } else {
-                    mException = new ImportedFileValidationException(
-                            fileImportValidator.getError());
-                    return null;
-                }
-            }
-            if (multiPageDocument.getDocuments().isEmpty()) {
-                mException = new ImportedFileValidationException("Intent did not contain images");
-                return null;
-            }
-            if (isCancelled()) {
-                return null;
-            }
-            return multiPageDocument;
-        }
-
-        @Override
-        protected void onPostExecute(final ImageMultiPageDocument multiPageDocument) {
-            if (multiPageDocument != null) {
-                mCallback.onDone(multiPageDocument);
-            } else if (mException != null) {
-                mCallback.onFailed(mException);
-            } else {
-                mCallback.onCancelled();
-            }
-        }
-    }
 }
